# API-Engine-distro/API-Engine-distro-data/src/main/resources/local/swagger.yaml
swagger: '2.0'
info:
  version: v1
  title: API Engine
basePath: /API-Engine-web/v1
...
schemes:
  - http
paths:
...
  '/organizations/{organizationId}/roles/{roleId}/{userId}':
    delete:
      tags:
        - organizations
      summary: Revoke Single Membership
      description: Revoke membership in a role.
      operationId: revoke
      parameters:
        - name: organizationId
          in: path
          required: true
          type: string
        - name: roleId
          in: path
          required: true
          type: string
        - name: userId
          in: path
          required: true
          type: string
      responses:
        '204':
          description: 'successful, no content'
...
  /roles:
    get:
      tags:
        - roles
      summary: List all Roles
      description: This endpoint lists all of the roles currently defined in apiman.
      operationId: list
      produces:
        - application/json
      parameters: []
      responses:
        '200':
          description: A list of roles.
          schema:
            type: array
            items:
              $ref: '#/definitions/SearchCriteriaBean'
    post:
      tags:
        - roles
      summary: Create Role
      description: Use this endpoint to create a new role.  A role consists of a set of permissions granted to a user when that user is given the role within the context of an organization.
      operationId: create
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: false
          schema:
            $ref: '#/definitions/NewRoleBean'
      responses:
        '200':
          description: The new created role.
          schema:
            $ref: '#/definitions/RoleBean'
  /roles/search:
    post:
      tags:
        - roles
      summary: Search for Roles
      description: 'This endpoint provides a way to search for roles. The search criteria is provided in the body of the request, including filters, order-by, and paging information.'
      operationId: search
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: false
          schema:
            $ref: '#/definitions/SearchCriteriaBean'
      responses:
        '200':
          description: A list of roles.
          schema:
            $ref: '#/definitions/SearchResultsBean'
  '/roles/{roleId}':
    get:
      tags:
        - roles
      summary: Get a Role by ID.
      description: Use this endpoint to retrieve information about a single Role by its ID.
      operationId: get
      produces:
        - application/json
      parameters:
        - name: roleId
          in: path
          required: true
          type: string
      responses:
        '200':
          description: A role.
          schema:
            $ref: '#/definitions/RoleBean'
    put:
      tags:
        - roles
      summary: Update a Role by ID
      description: Use this endpoint to update the information about an existing role.  The role is identified by its ID.
      operationId: update
      consumes:
        - application/json
      parameters:
        - name: roleId
          in: path
          required: true
          type: string
        - in: body
          name: body
          required: false
          schema:
            $ref: '#/definitions/UpdateRoleBean'
      responses:
        '204':
          description: 'successful, no content'
    delete:
      tags:
        - roles
      summary: Delete a Role by ID
      description: Use this endpoint to delete a role by its ID.
      operationId: delete
      parameters:
        - name: roleId
          in: path
          required: true
          type: string
      responses:
        '204':
          description: 'successful, no content'
...
definitions:
...
  PagingBean:
    type: object
    properties:
      page:
        type: integer
        format: int32
      pageSize:
        type: integer
        format: int32
...
  SearchResultsBean:
    type: object
    properties:
      beans:
        type: array
        items:
          type: object
      totalSize:
        type: integer
        format: int32
...
  SearchCriteriaBean:
    type: object
    properties:
      filters:
        type: array
        items:
          $ref: '#/definitions/SearchCriteriaFilterBean'
      orderBy:
        $ref: '#/definitions/OrderByBean'
      paging:
        $ref: '#/definitions/PagingBean'
...
  UpdateRoleBean:
    type: object
    properties:
      name:
        type: string
      description:
        type: string
      autoGrant:
        type: boolean
        default: false
      permissions:
        type: array
        uniqueItems: true
        items:
          type: string
          enum:
            - orgView
            - orgEdit
            - orgAdmin
            - svcView
            - svcEdit
            - svcAdmin
            - appView
            - appEdit
            - appAdmin
            - planView
            - planEdit
            - planAdmin
...
  NewRoleBean:
    type: object
    properties:
      name:
        type: string
      description:
        type: string
      autoGrant:
        type: boolean
        default: false
      permissions:
        type: array
        uniqueItems: true
        items:
          type: string
          enum:
            - orgView
            - orgEdit
            - orgAdmin
            - svcView
            - svcEdit
            - svcAdmin
            - appView
            - appEdit
            - appAdmin
            - planView
            - planEdit
            - planAdmin
...
  RoleBean:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      description:
        type: string
      createdBy:
        type: string
      createdOn:
        type: string
        format: date-time
      autoGrant:
        type: boolean
        default: false
      permissions:
        type: array
        uniqueItems: true
        items:
          type: string
          enum:
            - orgView
            - orgEdit
            - orgAdmin
            - svcView
            - svcEdit
            - svcAdmin
            - appView
            - appEdit
            - appAdmin
            - planView
            - planEdit
            - planAdmin
...
  OrderByBean:
    type: object
    properties:
      ascending:
        type: boolean
        default: false
      name:
        type: string
  SearchCriteriaFilterBean:
    type: object
    properties:
      name:
        type: string
      value:
        type: string
      operator:
        type: string
        enum:
          - bool_eq
          - eq
          - neq
          - gt
          - gte
          - lt
          - lte
          - like
...